Refactoring to service to be reusable
